+++ opentelemetry-etw-metrics/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## vNext
 
+- Bump opentelemetry and opentelemetry_sdk versions to 0.27
+- Bump opentelemetry-proto version to 0.27
+- Uses internal logging from `opentelemetry` crate, which routes internal logs
+  via `tracing`.
+
+## v0.5.0
+
+### Changed
+
+ - Bump opentelemetry and opentelemetry_sdk versions to 0.26
+ - Bump opentelemetry-proto version to 0.26
+ - Bump rust msrv to v1.71.1
+
 ## v0.4.0
 
 - Improved logging when ETW write fails due to size limit being hit.