+++ opentelemetry-user-events-logs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## vNext
 
+## v0.7.0
+
+### Changed
+
+- Bump opentelemetry and opentelemetry_sdk versions to 0.26
+
 ## v0.6.0
 
 ### Changed